move location to new file
